refactor: rename createBook to createBookFromFormData
- Create function createBookFromFormData in Book.ts
- Delete function createBook from main.ts
- Import function createBookFromFormData in main.ts
- Replace instances of use of createBook with createBookFromFormData
